fix datascience bot detection task
